requirements.txt
asgiref
bcrypt
certifi
cffi
cryptography
Django
django-cors-headers
django-seed
Faker
mysqlclient
pip
pycparser
PyJWT
PyMySQL
python-dateutil
pytz
setuptools
six
sqlparse
text-unidecode
wheel
